requirements.txt
torch==2.8.0
torchvision==0.23.0
faiss-cpu==1.12.0
numpy==1.26.4
scikit-learn==1.7.1
scipy==1.15.3
pandas==2.2.3
pytest==8.4.2
datasets==4.0.0 
transformers==4.56.0
sentence-transformers==5.1.0
huggingface-hub==0.34.4
llama-index-core==0.13.4
llama-index-embeddings-huggingface==0.6.0
llama-index-vector-stores-faiss==0.5.1
llama-index-instrumentation==0.4.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
tqdm==4.67.1